cs compiler: map IntType to int

+++ tests/machine/x/cs/append_builtin.cs
@@ -5,7 +5,7 @@ class Program
 {
     static void Main()
     {
-        List<long> a = new List<long> { 1, 2 };
-        Console.WriteLine(JsonSerializer.Serialize(new List<long>(a) { 3 }));
+        List<int> a = new List<int> { 1, 2 };
+        Console.WriteLine(JsonSerializer.Serialize(new List<int>(a) { 3 }));
     }
 }
