+++ tests/machine/x/cs/basic_compare.cs
@@ -4,8 +4,8 @@ class Program
 {
     static void Main()
     {
-        long a = (10 - 3);
-        long b = (2 + 2);
+        int a = (10 - 3);
+        int b = (2 + 2);
         Console.WriteLine(a);
         Console.WriteLine((a == 7));
         Console.WriteLine((b < 5));
