+++ tests/machine/x/cs/count_builtin.cs
@@ -6,6 +6,6 @@ class Program
 {
     static void Main()
     {
-        Console.WriteLine(Enumerable.Count(new List<long> { 1, 2, 3 }));
+        Console.WriteLine(Enumerable.Count(new List<int> { 1, 2, 3 }));
     }
 }
