+++ tests/machine/x/cs/for_map_collection.cs
@@ -1,13 +1,10 @@
 using System;
 using System.Collections.Generic;
 
-class Program
-{
-    static void Main()
-    {
-        Dictionary<string, long> m = new Dictionary<string, long> { { "a", 1 }, { "b", 2 } };
-        foreach (var k in m.Keys)
-        {
+class Program {
+    static void Main() {
+        Dictionary<string, int> m = new Dictionary<string, int> { { "a", 1 }, { "b", 2 } };
+        foreach (var k in m.Keys) {
             Console.WriteLine(k);
         }
     }
